Crate cookie_jar

Implementation of a RFC6265 compliant cookie store.

Modules

error
Errors produced by the cookie jar.

Structs

Attributes
The payload of the cookie including security requirements.

Builder
A builder for a cookie.

Cookie
This is the form that the cookie is represented in within the jar. It is formed by parsing the provided string into a cookie object.

Jar
A jar containing the cookies seen so far.

Pair
A decoded cookie name=value pair.

SetCookie
The SetCookie directive sent from the server.

Enums

Expires
Expiry time of a cookie.

Traits

Clock
Something that produces the current UTC time.