Struct cookie_jar::Attributes

pub struct Attributes { /* fields omitted */ }

The payload of the cookie including security requirements.

Methods

impl Attributes

pub fn pair(&self) -> &Pair

Get the (name, value) pair of a cookie.

pub fn secure(&self) -> bool

Check if the cookie requires a secure connection.

pub fn http_only(&self) -> bool

Check if the cookie should only be sent over http requests.

pub fn host_only(&self) -> bool

Check if the cookie is host-only.

pub fn expired(&self) -> bool

Check if the cookie has expired.

pub fn expired_since(&self, time: Tm) -> bool

Check if the cookie was expired after a given time.

pub fn expiry(&self) -> &Expires

Get the expiry of the cookie.