Struct cookie_jar::Cookie

pub struct Cookie { /* fields omitted */ }

This is the form that the cookie is represented in within the jar. It is formed by parsing the provided string into a cookie object.

Methods

impl Cookie

pub fn parse(set_cookie: &str, origin: &Url) -> Result<Cookie>

Parse a string cookie from a given origin.

pub fn host(&self) -> &Host

Get the domain or host the cookie applies to.

pub fn domain(&self) -> Option<&str>

Get the domain name associated with the cookie.

None if the cookie is host-only for an IP address.

pub fn path(&self) -> &str

Get the path the cookie applies to.